"multifile" + classes + my work of witches and interfaces.  - Problems with json
