+++ CardGame.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
+VisualStudioVersion = 17.7.34024.191
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleApp1", "ConsoleApp1.csproj", "{CAC8197D-EC47-4034-8882-0EB4936710D1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CardGame", "CardGame.csproj", "{D92B5EC4-4F11-4778-AAA2-498A7A9A3832}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CAC8197D-EC47-4034-8882-0EB4936710D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CAC8197D-EC47-4034-8882-0EB4936710D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CAC8197D-EC47-4034-8882-0EB4936710D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CAC8197D-EC47-4034-8882-0EB4936710D1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D92B5EC4-4F11-4778-AAA2-498A7A9A3832}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D92B5EC4-4F11-4778-AAA2-498A7A9A3832}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D92B5EC4-4F11-4778-AAA2-498A7A9A3832}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D92B5EC4-4F11-4778-AAA2-498A7A9A3832}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5EE1DAAA-B49D-4ACB-87C0-D92F0F4D44F9}
+		SolutionGuid = {55683E5F-C2E3-454A-BBB3-C16FC6371B30}
 	EndGlobalSection
 EndGlobal
